# jobportal-docs/apis/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/users": {
      "get": {
        "description": "Retrieve a list of all users in the system (Admin only)",
        "operationId": "AdminController_getAllUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get All Users",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/suspend/{userId}": {
      "patch": {
        "description": "Suspend a user account (Admin only)",
        "operationId": "AdminController_suspendUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID to suspend",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User suspended successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Suspend User",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/activate/{userId}": {
      "patch": {
        "description": "Activate a suspended user account (Admin only)",
        "operationId": "AdminController_activateUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID to activate",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User activated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Activate User",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/delete/{userId}": {
      "delete": {
        "description": "Permanently delete a user account (Admin only)",
        "operationId": "AdminController_deleteUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete User",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/report": {
      "get": {
        "description": "Get system usage statistics and analytics (Admin only)",
        "operationId": "AdminController_usageReport",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Report retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "totalUsers": 150,
                    "totalJobs": 45,
                    "totalApplications": 320,
                    "totalMessages": 1200,
                    "activeUsers": 140,
                    "suspendedUsers": 10,
                    "jobSeekers": 100,
                    "employers": 45,
                    "systemStats": {
                      "uptime": "5h 30m",
                      "memoryUsage": "45%",
                      "cpuUsage": "25%"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get Usage Report",
        "tags": [
          "Admin"
        ]
      }
    }
  },
  "info": {
    "title": "Job Portal API",
    "description": "A comprehensive API for managing jobs, applications, users, and messaging. Supports multiple user roles: JobSeeker, Employer, and Admin.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "JobSeeker",
              "Employer",
              "Admin"
            ]
          },
          "resumeUrl": {
            "type": "string",
            "example": "https://cloudinary.com/resume.pdf"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "address": {
            "type": "string",
            "example": "123 Main St, New York, NY"
          },
          "telephone": {
            "type": "string",
            "example": "+1-555-0199"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "email",
          "role",
          "name",
          "address",
          "telephone",
          "isActive"
        ]
      },
...
    }
  }
}